I am adding 6 files - Home.jsp, GetProductInformation.jsp, and ShowingProductInformation.jsp. The other 3 files are png of what these files look like during runtime.
On the Home.jsp, I have added 3 buttons & attached my test case (method, the form which takes the input for the method, and the page that shows the output of the method) to button 1 "Get Product Information". I also created the buttons for both of your test cases. Attach your form redirect to those buttons